RAG-Project/requirements.txt
streamlit
pymilvus
sentence-transformers
pytesseract
Pillow
docx
whisper
fitz
openai
